Create endpoint-security-detection
This query looks for process creation events, and then filters them in windows defender logs for the file 
types that exploits usually use and then filters it finally on the basis of the files types that run exploits. 

Note: Elevation of Privilege - CVE-2023-38175: This attack type uses Microsoft Windows Defender. A vulnerability in the MS Defender 
leads an attacker to perform an elevation of privilege to access the system.