Create web-application-detection
An improper header size limit leads to an incomplete POST request which then returns status code between 400 – 500. 
The incomplete POST requests trigger an error response which can have information from the previous request. 

Note: Information Disclosure (CVE-2023-46589): The attack which occurs with the ID – CVE-2023-46589 on Apache Tomcat is due to an 
improper input validation. The HTTP trailer headers were not parsed properly. This results in the treating of a single request as 
multiple requests when sent from behind a reverse proxy